Added a new model (of MVC) class for holding information about how the player performs during an individual level. It’s  avery simple class with just a bunch of integer properties for things like total score, number of apples obtained (of how many), etc. MyScene will update the numbers and pass an instance of the class to LevelEndView. Also replaced images for the buttons in LevelEndView and made some minor changes to LevelEndView to clean it up and make it consistent with the naming schemes in MyScene.

+++ MonkeySwing/JPMLevelInterpreter.h
@@ -10,6 +10,6 @@
 
 @interface JPMLevelInterpreter : NSObject
 
-+ (NSArray *)loadLevelFileNumber:(int)levelNumber;
++ (NSArray *)loadLevelFileNumber:(NSInteger)levelNumber;
 
 @end
